Added attack audio, Sword anim when attacking, renamed and retargeted names., and fixed open ended code in player state machine
Added sword swoosh animation and positioned it to always play on sword as well as sound for attack animation. Renamed player Sprite to reflect that it is the player sprite and called names in appropriate scripts. Fixed issue where there was an open ended section of code in player state machine causing a compiling error.